Sentinel/apps/server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
geoalchemy2
asyncpg
celery
redis
h3
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
